bump 20.08.16
- Fixed WP_Object_Cache::set() -> only write to disk if data change and expiry not 0.
- Fixed WP_Object_Cache::dc_precache_set -> only write to disk if data change.

bump 20.08.15

- Fixed precaching, invalid conditional for query string.
- Fixed cache maxttl, missing timestamp in cache meta.
- Fixed cache, flush user_meta group before login and after logout.
- Fixed micro optimization, before using regex functions.
- Fixed transient, remove all from db before activate our dropin.

+++ includes/src/Event.php
@@ -192,14 +192,6 @@ final class Event
                 $fm = time() + 120;
                 $ft = filemtime($fx);
 
-                if ($maxttl > 0 && $ft < $maxttl) {
-                    $this->plugin->unlink($fx, true);
-                    --$cnt;
-                    ++$collect->clean;
-                    ++$collect->maxttl_c;
-                    continue;
-                }
-
                 if ($fm >= $ft && (0 === $fs || 'dump_' === substr($fn, 0, 5))) {
                     $this->plugin->unlink($fx, true);
                     --$cnt;
@@ -209,13 +201,24 @@ final class Event
 
                 $data = $this->plugin->cache_get($fx);
                 if (false !== $data) {
-                    if (!empty($data['timeout']) && $fm >= (int) $data['timeout']) {
+                    if (!empty($data['timeout']) && $this->plugin->valid_timestamp($data['timeout']) && $fm >= (int) $data['timeout']) {
                         $this->plugin->unlink($fx, false);
                         unset($data);
                         --$cnt;
                         ++$collect->clean;
                         ++$collect->expired;
                         continue;
+                    }
+
+                    if (empty($data['timeout']) && !empty($data['timestamp']) && $this->plugin->valid_timestamp($data['timestamp']) && $maxttl > 0) {
+                        if ((int) $data['timestamp'] < $maxttl) {
+                            $this->plugin->unlink($fx, false);
+                            unset($data);
+                            --$cnt;
+                            ++$collect->clean;
+                            ++$collect->maxttl_c;
+                            continue;
+                        }
                     }
 
                     $bytestotal += \strlen(serialize($data));
@@ -227,6 +230,14 @@ final class Event
                     }
                 }
                 unset($data);
+
+                if ($maxttl > 0 && $ft < $maxttl) {
+                    $this->plugin->unlink($fx, true);
+                    --$cnt;
+                    ++$collect->clean;
+                    ++$collect->maxttl_c;
+                    continue;
+                }
 
                 if ($cnt >= $maxfile) {
                     $this->plugin->unlink($fx, true);
